Wizard support: Refactor `APIModel` via inheritance and create `src/model/api_model/model.ts`
