+++ src/model/api_model/model_wasm.ts
@@ -1,24 +1,16 @@
-import { ExtensionContext } from "../extensionContext";
-import { WasmWebviewPanel } from "../user_interface/wasmWebviewPanel";
-import { FSM } from "../model/fsm";
+import { ExtensionContext } from "../../extensionContext";
+import { WasmWebviewPanel } from "../../user_interface/wasmWebviewPanel";
+import { FSM } from "../../model/fsm";
 import * as vscode from 'vscode';
-import { ModelHelper } from "./utils/model_helper";
-import { WhammLiveInjection, WhammLiveResponse } from "./types";
-import { Cell } from "./utils/cell";
-import { Types } from "../whammServer";
-import { show_and_handle_error_response } from "../extensionListeners/documentChangesListener";
-import { Helper_sidebar_provider } from "../user_interface/sidebarProviderHelper";
-import { SvelteModel } from "./svelte_model";
+import { ModelHelper } from "../utils/model_helper";
+import { WhammLiveInjection, WhammLiveResponse } from "../types";
+import { Types } from "../../whammServer";
+import { show_and_handle_error_response } from "../../extensionListeners/documentChangesListener";
+import { Helper_sidebar_provider } from "../../user_interface/sidebarProviderHelper";
+import { APIModel } from "./model";
 
 // Class to store API responses [ MVC pattern's model ] 
-export class APIModel{
-    static whamm_file_changing: boolean = false;
-    api_response_setup_completed: boolean = false;
-    __api_response_out_of_date!: boolean;
-    // svelte side code updated or not
-    codemirror_code_updated!: boolean;
-
-    static whamm_cached_content: string;
+export class APIWasmModel extends APIModel{
     // will be null if wizard option chosen
     valid_wat_content!: string ;
     valid_wasm_content!: Uint8Array;
@@ -26,20 +18,14 @@ export class APIModel{
 
     // key is the wat line number and value is the whamm live injection at that line number
     wat_to_whamm_mapping: Map<number, WhammLiveInjection> = new Map();
-
-    injected_wat_content!: string;
-    jagged_array: (Cell|null)[][];
-
-    webview: WasmWebviewPanel;
     fsm_mappings: FSM | null;
     // funcID's change if whamm injects functions/ func imports
     // So, we need to a mapping to handle that 
     injected_fsm_mappings: FSM | null;
 
     constructor(webview: WasmWebviewPanel){
-        this.webview = webview;
+        super(webview);
         this.fsm_mappings = this.injected_fsm_mappings = null;
-        this.jagged_array = [];
     }
 
     async loadWatAndWasm(): Promise<boolean>{
@@ -127,7 +113,8 @@ export class APIModel{
                             this.wat_to_whamm_mapping = wat_to_whamm_mapping;
                         }
                         // update the jagged array
-                        this.update_jagged_array(this.whamm_live_response, file_contents);
+                        this.update_jagged_array(
+                            [...this.whamm_live_response.injecting_injections, ...this.whamm_live_response.other_injections], file_contents);
                         // success so return true
                         return true;
                     } else{
@@ -148,62 +135,5 @@ export class APIModel{
         }
         // unable to update since no whamm file
         return false;
-    }
-
-    async update_jagged_array(response: WhammLiveResponse, file_contents: string){
-        this.jagged_array = ModelHelper.create_jagged_array(file_contents);
-
-        // loop over every injection
-        // and for each injection's span
-        // inject the wamm-live-injection object into that cell of the jagged array
-        for (let injections of [response.injecting_injections, response.other_injections]){
-            for (let whamm_live_injection of injections){
-                if (whamm_live_injection.whamm_span === null) continue;
-
-                // Get all the (line,col) pairs for the injection's span
-                let line_col_values = ModelHelper.get_line_col_values(whamm_live_injection.whamm_span, this.jagged_array);
-                let span_size = line_col_values.length;
-                if (span_size <= 0) continue;
-
-                // loop over each (line,col) pair and add the injection to its cell value
-                for (let line_col of line_col_values){
-                    let row = line_col[0] -1;
-                    let col = line_col[1] -1;
-                    let cell = this.jagged_array[row][col];
-                    if (cell === null){
-                        this.jagged_array[line_col[0] -1][line_col[1] -1] = new Cell(whamm_live_injection, span_size);
-                    } else {
-                        cell.add(whamm_live_injection, span_size);
-                    }
-                }
-        }}
-    }
-
-    /* getter and setters */
-
-    get api_response_out_of_date(){
-        return this.__api_response_out_of_date;
-    }
-
-    // Updates the variables as well as notifies the svelte side by posting the messages
-    set api_response_out_of_date(value: boolean){
-        this.__api_response_out_of_date = value;
-        this.codemirror_code_updated = false;
-        SvelteModel.update_svelte_model(this.webview);
-    }
-
-    // set all models's api out of date and notify the related svelte side(s) only once!
-    static set_api_out_of_date(value: boolean){
-        for (let webview of WasmWebviewPanel.webviews){
-            webview.model.__api_response_out_of_date = value;
-            webview.model.codemirror_code_updated = false;
-        }
-        SvelteModel.update_svelte_models();
-    }
-
-    // file related helper static method(s)
-    static async loadFileAsString(path: string, context: vscode.ExtensionContext): Promise<string> {
-        const encoded = await vscode.workspace.fs.readFile(vscode.Uri.file(path));
-        return new TextDecoder('utf-8').decode(encoded);
     }
 }
